Different Version of Interlix I have been working on!
Basicly tells the user and welcomes them to "Interlix" Opens software then says they have installed a virus (Please note this is not a virus or malware or spyware or ransomware or a bug or a trojan horse or a computer virus or a scam). This is all a joke! (It's safe) 

Works Only On mac.